feat: AltTab-style unified tabbed shortcut editor (#74)
Rework the Shortcuts pane into a single AltTab-style editor: a tab bar lists
every panel-opening shortcut — the two core triggers (Apps, Windows) plus each
user-created scoped shortcut — with a "+" to add another. Selecting a tab shows
that shortcut's trigger recorder, its scope (scoped tabs only), and the full
set of per-shortcut options inline, all live-persisting as you change them.
Core tabs can't be removed; scoped tabs have a Remove button.

This replaces the previous two fixed "Switch apps / Switch windows" rows, the
separate scoped-shortcut list, and the modal "Customize…" sheet. Options now
edit in place (no draft/Done step), matching AltTab.

- ShortcutsEditorView: NSSegmentedControl tab bar + "+" and a detail panel
  rebuilt per selection (recorder + scope picker + inline form + remove).
- ShortcutOptionsFormView: the per-shortcut option grid, extracted from the old
  sheet and switched to write each change straight to Preferences.
- Removed ShortcutOptionsSheet and ScopedShortcutsListView (superseded).
- Search catalog: point the scoped-switch entry at the unified section.
